requirements.txt.txt
# Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Image Preprocessing
opencv-python==4.9.0.80
Pillow==10.1.0

# Data Handling and Visualization
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
pandas==2.0.3

# Evaluation Metrics
scikit-learn==1.3.0

# Streamlit Web App (optional, for UI)
streamlit==1.33.0

# For downloading from Google Drive (if needed)
gdown==5.1.0
